feat: :sparkles: working updateItem api; working on updating UI after successful operation
Integrated an ObjectId instance (instead of just a string), an atomic operator, and the returnDocument option of true with findOneAndUpdate to return the updated document to be used in updating the UI. Check appropriate properties of DAO result objects to return the appropriate status codes. Form closes & user is notified that the item was updated successfully, if status code 200 is returned.